% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2metid.R
\name{convert2metid}
\alias{convert2metid}
\title{Convert data (list, from read_msp_data function) to metID format database}
\usage{
convert2metid(data, path = ".", threads = 5, Submitter = "Xiaotao Shen")
}
\arguments{
\item{data}{list, from read_msp_data function}

\item{path}{Default is .}

\item{threads}{threads}

\item{Submitter}{Default is "Xiaotao Shen"}
}
\value{
metid database class
}
\description{
Convert data (list, from read_msp_data function) to metID format database
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
