% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_BIGG.R
\name{convert_bigg_universal2metid}
\alias{convert_bigg_universal2metid}
\title{Convert BIGG universal metabolites (data.frame,
from read_bigg_universal_metabolite)
to metID format database}
\usage{
convert_bigg_universal2metid(data, path = ".", threads = 5)
}
\arguments{
\item{data}{data.frame, from read_bigg_universal_metabolite function}

\item{path}{Default is .}

\item{threads}{threads}
}
\value{
metid database class
}
\description{
Convert BIGG universal metabolites (data.frame,
from read_bigg_universal_metabolite)
to metID format database
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
