% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_CHEBI.R
\name{convert_chebi2metid}
\alias{convert_chebi2metid}
\title{Convert BIGG universal metabolites (data.frame,
from read_chebi_metabolite)
to metID format database}
\usage{
convert_chebi2metid(data, path = ".", threads = 5)
}
\arguments{
\item{data}{data.frame, from read_chebi_metabolite function}

\item{path}{Default is .}

\item{threads}{threads}
}
\value{
metid database class
}
\description{
Convert BIGG universal metabolites (data.frame,
from read_chebi_metabolite)
to metID format database
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
