% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_FOODB.R
\name{convert_foodb2metid}
\alias{convert_foodb2metid}
\title{Convert FOODB compound data (list,
from download_foodb_compound function)
to metID format database}
\usage{
convert_foodb2metid(data, path = ".", threads = 5)
}
\arguments{
\item{data}{data.frame, from read_foodb_compound function.}

\item{path}{Default is .}

\item{threads}{threads}
}
\value{
metid database class
}
\description{
Convert FOODB compound data (list,
from download_foodb_compound function)
to metID format database
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
