% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_GNPS.R
\name{convert_gnps2metid}
\alias{convert_gnps2metid}
\title{Convert GNPS data (list) to metID format database}
\usage{
convert_gnps2metid(data, path = ".", threads = 5)
}
\arguments{
\item{data}{list, from read_msp_data function}

\item{path}{Default is .}

\item{threads}{threads}
}
\value{
metid database class
}
\description{
Convert GNPS data (list) to metID format database
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
