% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_HMDB.R
\name{convert_hmdb2metid}
\alias{convert_hmdb2metid}
\title{Convert HMDB data (MS1 or MS2) to metID format database}
\usage{
convert_hmdb2metid(data, path = ".", threads = 5, ms1_or_ms2 = c("ms1", "ms2"))
}
\arguments{
\item{data}{for MS1, it is a data.from from read_xml_data_hmdb.}

\item{path}{Default is .}

\item{threads}{threads}

\item{ms1_or_ms2}{MS1 or MS2 data.}
}
\value{
metid database class
}
\description{
Convert HMDB data (MS1 or MS2) to metID format database
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
