% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16_KEGG.R
\name{convert_kegg2metid}
\alias{convert_kegg2metid}
\title{Convert KEGG compound/drug data (data.frame,
from read_kegg_compound or read_kegg_drug function)
to metID format database}
\usage{
convert_kegg2metid(data, path = ".", threads = 5)
}
\arguments{
\item{data}{data.frame, from read_kegg_compound or read_kegg_drug function}

\item{path}{Default is .}

\item{threads}{threads}
}
\value{
metid database class
}
\description{
KEGG compound/drug data (data.frame,
from read_kegg_compound or read_kegg_drug function)
to metID format database
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
