% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16_KEGG_pathway.R
\name{convert_kegg2metpath}
\alias{convert_kegg2metpath}
\title{Convert KEGG pathway data (list,
from read_kegg_pathway function)
to metpath format database}
\usage{
convert_kegg2metpath(data, path = ".", threads = 5)
}
\arguments{
\item{data}{list, from read_kegg_pathway function}

\item{path}{Default is .}

\item{threads}{threads}
}
\value{
metpath pathway_database class
}
\description{
Convert KEGG pathway data (list,
from read_kegg_pathway function)
to metpath format database
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
