% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIPIDBANK.R
\name{convert_lipidbank2metid}
\alias{convert_lipidbank2metid}
\title{Convert lipid bank data (data.frame) to metID format database}
\usage{
convert_lipidbank2metid(data)
}
\arguments{
\item{data}{data.frame}
}
\value{
Downloaded files.
}
\description{
Convert lipid bank data (data.frame) to metID format database
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
