% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13_LIPIDMAPS.R
\name{convert_lipidmaps2metid}
\alias{convert_lipidmaps2metid}
\title{Convert LIPIDMAPS compound data (list,
from read_sdf_data_lipidmaps function)
to metID format database}
\usage{
convert_lipidmaps2metid(data, path = ".", threads = 5)
}
\arguments{
\item{data}{list, from read_sdf_data_lipidmaps function}

\item{path}{Default is .}

\item{threads}{threads}
}
\value{
metid database class
}
\description{
Convert LIPIDMAPS compound data (list,
from read_sdf_data_lipidmaps function)
to metID format database
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
