% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_MASSBANK.R
\name{convert_massbank2metid}
\alias{convert_massbank2metid}
\title{Convert MassBank data (list) to metID format database}
\usage{
convert_massbank2metid(
  data,
  source = c("riken", "nist"),
  path = ".",
  threads = 5
)
}
\arguments{
\item{data}{list, from read_msp_data function}

\item{source}{riken or nist}

\item{path}{Default is .}

\item{threads}{threads}
}
\value{
metid database class
}
\description{
Convert MassBank data (list) to metID format database
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
