% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_SMPDB_PATHWAY.R
\name{convert_smpdb2metpath}
\alias{convert_smpdb2metpath}
\title{Convert smpdb pathway data (list,
from read_smpdb_pathway function)
to metpath format database}
\usage{
convert_smpdb2metpath(data, path = ".", threads = 5)
}
\arguments{
\item{data}{list, from read_smpdb_pathway function}

\item{path}{Default is .}

\item{threads}{threads}
}
\value{
metpath pathway_database class
}
\description{
Convert smpdb pathway data (list,
from read_smpdb_pathway function)
to metpath format database
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
