% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_species2source}
\alias{convert_species2source}
\title{Convert species to source}
\usage{
convert_species2source(x, match_table)
}
\arguments{
\item{x}{source}

\item{match_table}{match_table}
}
\value{
A data.frame
}
\description{
Convert species to source
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
