% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIGG.R
\name{down_bigg_model}
\alias{down_bigg_model}
\title{Download BIGG model data}
\usage{
down_bigg_model(model_id = "iND750")
}
\arguments{
\item{model_id}{model_id, for example, iND750}
}
\value{
version.
}
\description{
Download BIGG model data
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
