% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_BIGG.R
\name{download_bigg_universal_metabolite}
\alias{download_bigg_universal_metabolite}
\title{Download BIGG universal metabolites}
\usage{
download_bigg_universal_metabolite(path = ".", sleep = 1)
}
\arguments{
\item{path}{Default is .}

\item{sleep}{Default is 1 second.}
}
\value{
bigg_universal_metabolite database.
}
\description{
Download BIGG universal metabolites
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
