% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHEBI.R
\name{download_chebi_compound}
\alias{download_chebi_compound}
\title{Download the CHEBI compound database}
\usage{
download_chebi_compound(
  url = "https://ftp.ebi.ac.uk/pub/databases/chebi/Flat_file_tab_delimited/",
  path = "."
)
}
\arguments{
\item{url}{Default is "https://ftp.ebi.ac.uk/pub/databases/chebi/Flat_file_tab_delimited/".}

\item{path}{Default is ..}
}
\value{
Downloaded files.
}
\description{
Download the CHEBI compound database
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
