% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_GNPS.R
\name{download_gnps_spectral_library}
\alias{download_gnps_spectral_library}
\title{Download GNPS MS2 database (msp)}
\usage{
download_gnps_spectral_library(
  url = "https://gnps-external.ucsd.edu/gnpslibrary",
  gnps_library = "HMDB",
  path = "."
)
}
\arguments{
\item{url}{Default is "https://gnps-external.ucsd.edu/gnpslibrary".}

\item{gnps_library}{See here: https://gnps-external.ucsd.edu/gnpslibrary.}

\item{path}{Default is .}
}
\value{
Downloaded files.
}
\description{
Download GNPS MS2 database (msp)
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
