% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16_KEGG.R
\name{download_kegg_compound}
\alias{download_kegg_compound}
\title{Download KEGG compound data}
\usage{
download_kegg_compound(path = ".", sleep = 1)
}
\arguments{
\item{path}{Default is .}

\item{sleep}{Default is 1 second.}
}
\value{
KEGG compound database, rda format.
}
\description{
Download KEGG compound data
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
