% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16_KEGG.R
\name{download_kegg_drug}
\alias{download_kegg_drug}
\title{Download KEGG drug data}
\usage{
download_kegg_drug(path = ".", sleep = 1)
}
\arguments{
\item{path}{Default is .}

\item{sleep}{Default is 1 second.}
}
\value{
KEGG drug database, rda format.
}
\description{
Download KEGG drug data
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
