% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16_KEGG_pathway.R
\name{download_kegg_pathway}
\alias{download_kegg_pathway}
\title{Download KEGG pathway data}
\usage{
download_kegg_pathway(path = ".", sleep = 1, organism = "hsa")
}
\arguments{
\item{path}{Default is .}

\item{sleep}{Default is 1 second.}

\item{organism}{organism. see https://www.genome.jp/kegg/pathway.html}
}
\value{
KEGG pathway database, rda format.
}
\description{
Download KEGG pathway data
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
