% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16_KEGG.R
\name{download_kegg_reaction}
\alias{download_kegg_reaction}
\title{Download KEGG reaction database}
\usage{
download_kegg_reaction(path = ".", sleep = 1)
}
\arguments{
\item{path}{Default is .}

\item{sleep}{Default is 1 second.}
}
\value{
KEGG reaction database, R format.
}
\description{
Download KEGG reaction database
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
