% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9_LIPIDBANK.R
\name{download_lipidbank_lipid_class}
\alias{download_lipidbank_lipid_class}
\title{Download lIPIDBANK database}
\usage{
download_lipidbank_lipid_class(
  url = "https://lipidbank.jp/download/",
  lipid_class = c("All data", "Acylglycerol", "Bile Acid", "Fatty acid",
    "Long chain alcohol", "Long chain aldehyde", "Long chain base and Ceramide",
    "Eicosanoid", "Ether type lipid", "Carotenoid", "Coenzyme Q", "Vitamin A",
    "Vitamin D", "Vitamin E", "Vitamin F", "Vitamin K", "Glycosphingolipid",
    "Glycoglycerolipid and others", "Isoprenoid", "Lipid peroxide", "Lipoamino acid",
    "Lipopolysaccharide", "Lipoprotein", "Mycolic acid", "Glycerophospholipid",
    "Sphingophospholipid", "Steroid", "Wax"),
  path = "."
)
}
\arguments{
\item{url}{Default is "https://lipidbank.jp/download/".}

\item{lipid_class}{See here: https://lipidbank.jp/index.html.}

\item{path}{Default is .}
}
\value{
Downloaded files.
}
\description{
Download lIPIDBANK database
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
