% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13_LIPIDMAPS.R
\name{download_lipidmaps_lipid}
\alias{download_lipidmaps_lipid}
\title{Download LIPIDMAPS data}
\usage{
download_lipidmaps_lipid(path = ".")
}
\arguments{
\item{path}{Default is .}
}
\value{
version.
}
\description{
Download LIPIDMAPS data
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
