% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_MASSBANK.R
\name{download_massbank_compound}
\alias{download_massbank_compound}
\title{Download the MassBank compound database}
\usage{
download_massbank_compound(
  url = "https://github.com/MassBank/MassBank-data/releases/download/2021.12",
  source = c("nist", "riken"),
  path = "."
)
}
\arguments{
\item{url}{Default is "https://github.com/MassBank/MassBank-data/releases/download/2021.12".}

\item{source}{nist or riken.}

\item{path}{Default is ..}
}
\value{
Downloaded files.
}
\description{
Download the MassBank compound database
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
