% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15_PUBCHEM.R
\name{download_pubchem_compound}
\alias{download_pubchem_compound}
\title{Request the information of one compound from PubChem}
\usage{
download_pubchem_compound(part_index = as.character(1:329), path = ".")
}
\arguments{
\item{part_index}{Which part you want to download. See here:
https://ftp.ncbi.nlm.nih.gov/pubchem/Compound/CURRENT-Full/XML/}

\item{path}{default is .}
}
\value{
xml data
}
\description{
Request the information of one compound from PubChem
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
