% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_SMPDB_PATHWAY.R
\name{download_smpdb_pathway}
\alias{download_smpdb_pathway}
\title{Download SMPDB pathway data}
\usage{
download_smpdb_pathway(path = ".")
}
\arguments{
\item{path}{Default is .}
}
\value{
SMPDB pathway database, csv format.
}
\description{
Download SMPDB pathway data
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
