% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_words_similarity.R
\name{get_words_similarity}
\alias{get_words_similarity}
\title{Get the similarity between two words}
\usage{
get_words_similarity(
  word1 = "(2S)-2-Amino-3-(1-methyl-1H-imidazol-4-yl)propanoic acid",
  word2 = "(2S)-2-Amino-3-(1-methyl-1H-imidazol-4-yl)propanoate",
  ...
)
}
\arguments{
\item{word1}{word1.}

\item{word2}{word2.}

\item{...}{Other arguments for stringdist::stringsim()}
}
\value{
A simility score.
}
\description{
Get the similarity between two words
}
\examples{
get_words_similarity(word1 = "(2S)-2-Amino-3-(1-methyl-1H-imidazol-4-yl)propanoic acid",
word2 = "(2S)-2-Amino-3-(1-methyl-1H-imidazol-4-yl)propanoate")
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
