% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massdatabase_logo.R
\name{massdatabase_logo}
\alias{massdatabase_logo}
\title{Show the base information of massdatabase pacakge}
\usage{
massdatabase_logo()
}
\value{
A ASCII log of massdatabase
}
\description{
Show the base information of massdatabase pacakge.
}
\examples{
massdatabase_logo()
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
