% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{match_table}
\alias{match_table}
\title{match table}
\format{
A data frame.
}
\usage{
match_table
}
\description{
match table
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
\keyword{datasets}
