% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/19_REACTOME.R
\name{parse_reactome_pathway}
\alias{parse_reactome_pathway}
\title{Parse Reactome Pathway SBML File}
\usage{
parse_reactome_pathway(file_name)
}
\arguments{
\item{file_name}{Character. The file path of the SBML file.}
}
\value{
A list containing:
  \item{pathway_name}{Character. Name of the pathway.}
  \item{pathway_id}{Character. Reactome pathway ID.}
  \item{pathway_description}{Character. Text description of the pathway.}
  \item{component_info}{Data frame containing metabolites (CHEBI) and proteins (UniProt) involved in the pathway.}
}
\description{
This function extracts pathway details, including description, metabolites, and proteins from an SBML file.
}
\examples{
# Download a sample SBML file from Reactome and parse it
pathway_info <- request_reactome_pathway("R-HSA-5652084")
print(pathway_info$pathway_name)
print(pathway_info$component_info)
}
\author{
Xiaotao Shen (\email{xiaotao.shen@outlook.com})
}
