% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/19_REACTOME.R
\name{parse_reactome_reaction}
\alias{parse_reactome_reaction}
\title{Parse the sbml reaction data from Reactome}
\usage{
parse_reactome_reaction(file_name)
}
\arguments{
\item{file_name}{file name of the data.}
}
\value{
A data frame or list.
}
\description{
Parse the sbml reaction data from Reactome
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
