% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_BIGG.R
\name{read_bigg_model}
\alias{read_bigg_model}
\title{Read the BIGG model from download_bigg_model function}
\usage{
read_bigg_model(path = ".", model = "iND750")
}
\arguments{
\item{path}{Default is .. Should be same with
download_bigg_model function.}

\item{model}{model}
}
\value{
A data frame
}
\description{
Read the BIGG model from download_bigg_model function
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
