% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_BIGG.R
\name{read_bigg_universal_metabolite}
\alias{read_bigg_universal_metabolite}
\title{Read the BIGG universal metabolite database from
download_bigg_universal_metabolite function}
\usage{
read_bigg_universal_metabolite(path = ".")
}
\arguments{
\item{path}{Default is .. Should be same with
download_bigg_universal_metabolite function.}
}
\value{
A data frame
}
\description{
Read the BIGG universal metabolite database from
download_bigg_universal_metabolite function
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
