% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHEBI.R
\name{read_chebi_compound}
\alias{read_chebi_compound}
\title{Read the CHEBI compound database from download_chebi_compound function}
\usage{
read_chebi_compound(path = ".")
}
\arguments{
\item{path}{Default is .. Should be same with download_chebi_compound function.}
}
\value{
A data frame
}
\description{
Read the CHEBI compound database from download_chebi_compound function
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
