% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_FOODB.R
\name{read_foodb_compound}
\alias{read_foodb_compound}
\title{Read the FOODB compound database from download_foodb_compound function}
\usage{
read_foodb_compound(path = ".")
}
\arguments{
\item{path}{Default is .. Should be same with download_foodb_compound function.}
}
\value{
A data frame
}
\description{
Read the FOODB compound database from download_foodb_compound function
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
