% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_utils.R
\name{read_gpml}
\alias{read_gpml}
\title{Read GPML format data from wikipathway database}
\usage{
read_gpml(file, only_remain_metabolites = TRUE)
}
\arguments{
\item{file}{file name of GPML data}

\item{only_remain_metabolites}{only remain metabolites inforamtion or not.}
}
\value{
A data.frame
}
\description{
Read GPML format data from wikipathway database
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
