% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16_KEGG.R
\name{read_kegg_compound}
\alias{read_kegg_compound}
\title{Read the KEGG compound database from download_kegg_compound function}
\usage{
read_kegg_compound(path = ".")
}
\arguments{
\item{path}{Default is .. Should be same with download_kegg_compound function.}
}
\value{
A data frame
}
\description{
Read the KEGG compound database from download_kegg_compound function
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
