% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16_KEGG.R
\name{read_kegg_drug}
\alias{read_kegg_drug}
\title{Read the KEGG drug database from download_kegg_drug function}
\usage{
read_kegg_drug(path = ".")
}
\arguments{
\item{path}{Default is .. Should be same with download_kegg_drug function.}
}
\value{
A data frame
}
\description{
Read the KEGG drug database from download_kegg_drug function
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
