% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16_KEGG_pathway.R
\name{read_kegg_pathway}
\alias{read_kegg_pathway}
\title{Read the KEGG pathway database from download_kegg_pathway function}
\usage{
read_kegg_pathway(path = ".")
}
\arguments{
\item{path}{Default is .. Should be same with download_kegg_pathway function.}
}
\value{
A data frame
}
\description{
Read the KEGG pathway database from download_kegg_pathway function
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
