% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_msp_data.R
\name{read_msp_data}
\alias{read_msp_data}
\title{read_msp_data}
\usage{
read_msp_data(
  file,
  source = c("massbank", "mona", "gnps", "nist"),
  threads = 5
)
}
\arguments{
\item{file}{The vector of names of ms2 files. MS2 file must be msp format.}

\item{source}{massbank, mona.}

\item{threads}{threads}
}
\value{
Return ms2 data. This is a list.
}
\description{
Read MSP data
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
