% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13_LIPIDMAPS.R
\name{read_sdf_data_lipidmaps}
\alias{read_sdf_data_lipidmaps}
\title{Read the sdf database from download_lipidmaps_lipid function}
\usage{
read_sdf_data_lipidmaps(file, path = ".")
}
\arguments{
\item{file}{should be sdf format}

\item{path}{Default is .. Should be same with download_lipidmaps_lipid function.}
}
\value{
A data frame
}
\description{
Read the sdf database from download_lipidmaps_lipid function
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
