% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_SMPDB_PATHWAY.R
\name{read_smpdb_pathway}
\alias{read_smpdb_pathway}
\title{Read the SMPDB pathway database from download_smpdb_pathway function}
\usage{
read_smpdb_pathway(
  path = ".",
  only_primarity_pathway = TRUE,
  remain_disease_pathway = TRUE,
  remain_drug_pathway = TRUE
)
}
\arguments{
\item{path}{Default is .. Should be same with download_smpdb_pathway function.}

\item{only_primarity_pathway}{Only remain primary pathway?}

\item{remain_disease_pathway}{remain disease pathway?}

\item{remain_drug_pathway}{remain drug pathway?}
}
\value{
A data frame
}
\description{
Read the SMPDB pathway database from download_smpdb_pathway function
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
