% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_xml_data.R
\name{read_xml_data}
\alias{read_xml_data}
\title{read_xml_data}
\usage{
read_xml_data(
  file,
  source = c("hmdb"),
  ms1_or_ms2 = c("ms1", "ms2"),
  path = ".",
  threads = 5
)
}
\arguments{
\item{file}{The vector of names of files.}

\item{source}{hmdb}

\item{ms1_or_ms2}{ms1 or ms2 data.}

\item{path}{where is the file(s).}

\item{threads}{threads}
}
\value{
Return ms2 data. This is a list.
}
\description{
Read xml data
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
