% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_xml_data.R
\name{read_xml_data_hmdb}
\alias{read_xml_data_hmdb}
\title{read_xml_data_hmdb}
\usage{
read_xml_data_hmdb(file, threads = 5, path = ".", ms1_or_ms2 = c("ms1", "ms2"))
}
\arguments{
\item{file}{The file name(s).}

\item{threads}{threads}

\item{path}{where is/are the file(s)}

\item{ms1_or_ms2}{MS1 or MS2 data.}
}
\value{
Return ms2 data. This is a list.
}
\description{
Read xml data from HMDB
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
