% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIGG.R
\name{request_bigg_model_info}
\alias{request_bigg_model_info}
\title{Request BIGG model information}
\usage{
request_bigg_model_info(url = "http://bigg.ucsd.edu/api/v2/models")
}
\arguments{
\item{url}{url http://bigg.ucsd.edu/api/v2/models}
}
\value{
model inforamtion
}
\description{
Request BIGG model information
}
\examples{
x <- request_bigg_model_info()
head(x)
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
