% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_BIGG.R
\name{request_bigg_reaction_info}
\alias{request_bigg_reaction_info}
\title{Request BIGG reaction information for specific species (model)}
\usage{
request_bigg_reaction_info(model_bigg_id = "iAB_RBC_283")
}
\arguments{
\item{model_bigg_id}{please use the model bigg id using the
request_bigg_model_info function. Default is iAB_RBC_283 (human)}
}
\value{
reaction information
}
\description{
Request BIGG reaction information for specific species
}
\examples{
x <- request_bigg_universal_reaction_info()
dim(x)
head(x)
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
