% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_BIGG.R
\name{request_bigg_universal_metabolite}
\alias{request_bigg_universal_metabolite}
\title{Request BIGG metabolite}
\usage{
request_bigg_universal_metabolite(
  url = "http://bigg.ucsd.edu/api/v2/universal/metabolites",
  metabolite_id = "g3p",
  return_form = c("list", "data.frame")
)
}
\arguments{
\item{url}{url http://bigg.ucsd.edu/api/v2/universal/metabolites}

\item{metabolite_id}{metabolite_id, for example, g3p}

\item{return_form}{data.frame or list.}
}
\value{
universal metabolite inforamtion
}
\description{
Request BIGG metabolite
}
\examples{
x <- request_bigg_universal_metabolite(metabolite_id = "g3p", return_form = "list")
x
y <- request_bigg_universal_metabolite(metabolite_id = "g3p", return_form = "data.frame")
y
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
