% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_BIGG.R
\name{request_bigg_universal_reaction}
\alias{request_bigg_universal_reaction}
\title{Request BIGG reaction}
\usage{
request_bigg_universal_reaction(
  url = "http://bigg.ucsd.edu/api/v2/universal/reactions",
  reaction_id = "ADA",
  return_form = c("list", "data.frame")
)
}
\arguments{
\item{url}{url http://bigg.ucsd.edu/api/v2/universal/reactions}

\item{reaction_id}{reaction_id, for example, ADA}

\item{return_form}{data.frame or list.}
}
\value{
universal metabolite inforamtion
}
\description{
Request BIGG reaction
}
\examples{
x <- request_bigg_universal_reaction(reaction_id = "ADA", return_form = "list")
x
y <- request_bigg_universal_reaction(reaction_id = "ADA", return_form = "data.frame")
y
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
