% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_BIGG.R
\name{request_bigg_universal_reaction_info}
\alias{request_bigg_universal_reaction_info}
\title{Request BIGG universal reaction information}
\usage{
request_bigg_universal_reaction_info(
  url = "http://bigg.ucsd.edu/api/v2/universal/reactions"
)
}
\arguments{
\item{url}{url http://bigg.ucsd.edu/api/v2/universal/reactions}
}
\value{
universal metabolite inforamtion
}
\description{
Request BIGG universal reaction information
}
\examples{
x <- request_bigg_universal_reaction_info()
dim(x)
head(x)
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
