% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_CHEBI.R
\name{request_chebi_compound}
\alias{request_chebi_compound}
\title{Request CHEBI compound}
\usage{
request_chebi_compound(
  url = "https://www.ebi.ac.uk/chebi/searchId.do?chebiId=",
  compound_id = "CHEBI:18358"
)
}
\arguments{
\item{url}{url https://www.ebi.ac.uk/chebi/searchId.do?chebiId=}

\item{compound_id}{compound_id, for example, CHEBI:18358}
}
\value{
Compound information (list)
}
\description{
Request CHEBI compound
}
\examples{
request_chebi_compound(compound_id = "CHEBI:18358")
request_chebi_compound(compound_id = "CHEBI:18377")
request_chebi_compound(compound_id = "CHEBI:18399")
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
