% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodb.R
\name{request_foodb_compound}
\alias{request_foodb_compound}
\title{Request one specific the compound information in FoodB}
\usage{
request_foodb_compound(
  url = "https://foodb.ca/compounds",
  compound_id = "FDB000004",
  return_form = c("list", "data.frame")
)
}
\arguments{
\item{url}{Default is "https://foodb.ca/compounds".}

\item{compound_id}{compound id. For example, FDB000004.}

\item{return_form}{data.frame or list.}
}
\value{
A data frame or list.
}
\description{
Request one specific the compound information in FoodB
}
\examples{
x = request_foodb_compound( compound_id = "FDB000004", return_form = "list")
x
y = request_foodb_compound( compound_id = "FDB000004", return_form = "data.frame")

}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
