% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_FOODB.R
\name{request_foodb_compound_info}
\alias{request_foodb_compound_info}
\title{Request all the food information in FoodB}
\usage{
request_foodb_compound_info(
 
    url = c("https://raw.githubusercontent.com/jaspershen/databases/main/data/foodb_compound_info.rda",
    "https://foodb.ca/downloads", "https://foodb.ca/compounds"),
  sleep = 1,
  pages = 1:2838
)
}
\arguments{
\item{url}{Default is "https://foodb.ca/downloads".}

\item{sleep}{sleep default is 1 second.}

\item{pages}{default is from 1:2838}
}
\value{
A data frame.
}
\description{
Request all the food information in FoodB
}
\examples{
x = request_foodb_compound_info()
head(x)
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
