% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FOODB.R
\name{request_foodb_compound_ms2}
\alias{request_foodb_compound_ms2}
\title{Request MS2 spectra of one compound in FoodB}
\usage{
request_foodb_compound_ms2(compound_id = "FDB000013")
}
\arguments{
\item{compound_id}{compound id. For example, FDB000004.}
}
\value{
A data frame
}
\description{
Request MS2 spectra of one compound in FoodB
}
\examples{
x = request_foodb_compound_ms2(compound_id = "FDB000004")
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
