% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_FOODB.R
\name{request_foodb_food_info}
\alias{request_foodb_food_info}
\title{Request all the food information in FoodB}
\usage{
request_foodb_food_info(
 
    url = c("https://raw.githubusercontent.com/jaspershen/databases/main/data/foodb_compound_info.rda",
    "https://foodb.ca/downloads")
)
}
\arguments{
\item{url}{Default is "https://foodb.ca/downloads".}
}
\value{
A data frame.
}
\description{
Request all the food information in FoodB
}
\examples{
x = request_foodb_food_info()
head(x)
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
