% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMDB.R
\name{request_hmdb_metabolite}
\alias{request_hmdb_metabolite}
\title{Request one specific the metabolite information in HMDB}
\usage{
request_hmdb_metabolite(
  url = "https://hmdb.ca/metabolites",
  metabolite_id = "HMDB0000001",
  return_form = c("list", "data.frame")
)
}
\arguments{
\item{url}{Default is "https://hmdb.ca/metabolites".}

\item{metabolite_id}{metabolite id. For example, HMDB0000001}

\item{return_form}{data.frame or list.}
}
\value{
A data frame or list.
}
\description{
Request one specific the metabolite information in HMDB
}
\examples{
x = request_hmdb_metabolite(metabolite_id = "HMDB0000001", return_form = "list")
x[1:2]
y = request_hmdb_metabolite(metabolite_id = "HMDB0000001", return_form = "data.frame")
head(y)
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
