% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_HMDB.R
\name{request_hmdb_reaction}
\alias{request_hmdb_reaction}
\title{Request one specific reaction information in HMDB}
\usage{
request_hmdb_reaction(url = "https://hmdb.ca/reactions", reaction_id = "241")
}
\arguments{
\item{url}{Default is "https://hmdb.ca/reactions".}

\item{reaction_id}{reaction id. For example, 241}
}
\value{
A data frame .
}
\description{
Request one specific reaction information in HMDB
}
\examples{
request_hmdb_reaction(reaction_id = "1")
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
