% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16_KEGG.R
\name{request_kegg_compound}
\alias{request_kegg_compound}
\title{Request one specific compound information in KEGG}
\usage{
request_kegg_compound(
  compound_id = "C02886",
  return_form = c("list", "data.frame")
)
}
\arguments{
\item{compound_id}{compound id. For example, C02886}

\item{return_form}{data.frame or list.}
}
\value{
A data frame or list.
}
\description{
Request one specific compound information in KEGG
}
\examples{
x = request_kegg_compound(compound_id = "C02886", return_form = "list")
x[1:2]
y = request_kegg_compound(compound_id = "C02886", return_form = "data.frame")
head(y)
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
