% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16_KEGG_pathway.R
\name{request_kegg_pathway}
\alias{request_kegg_pathway}
\title{Request one specific pathway information in KEGG}
\usage{
request_kegg_pathway(pathway_id = "hsa00010")
}
\arguments{
\item{pathway_id}{pathway id. For example, hsa00010}
}
\description{
Request one specific pathway information in KEGG
}
\examples{
x = request_kegg_pathway(pathway_id = "hsa00010")
x[1:2]

}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
