% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16_KEGG_pathway.R
\name{request_kegg_pathway_info}
\alias{request_kegg_pathway_info}
\title{Request all pathway information in KEGG}
\usage{
request_kegg_pathway_info(organism = "hsa")
}
\arguments{
\item{organism}{organism. see https://www.genome.jp/kegg/pathway.html}
}
\value{
A data frame.
}
\description{
Request all pathway information in KEGG
}
\examples{
head(request_kegg_pathway_info, 3)
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
