% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16_KEGG.R
\name{request_kegg_rclass}
\alias{request_kegg_rclass}
\title{Request one specific reaction class information in KEGG}
\usage{
request_kegg_rclass(rclass_id = "RC00001")
}
\arguments{
\item{rclass_id}{reaction class id. For example, RC00001}
}
\value{
A data frame or list.
}
\description{
Request one specific reaction class information in KEGG
}
\examples{
x = request_kegg_rclass("RC00001")
x
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
