% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16_KEGG.R
\name{request_kegg_rclass_info}
\alias{request_kegg_rclass_info}
\title{Request all reaction class information in KEGG}
\usage{
request_kegg_rclass_info()
}
\value{
A data frame.
}
\description{
Request all reaction class information in KEGG
}
\examples{
head(request_kegg_rclass_info(), 3)
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
