% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16_KEGG.R
\name{request_kegg_reaction}
\alias{request_kegg_reaction}
\title{Request one specific reaction information in KEGG}
\usage{
request_kegg_reaction(reaction_id = "R00001")
}
\arguments{
\item{reaction_id}{reaction id. For example, R00001}
}
\value{
A data frame or list.
}
\description{
Request one reaction compound information in KEGG
}
\examples{
x = request_kegg_reaction("R00001")
x
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
