% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9_LIPIDBANK.R
\name{request_lipidbank_lipid_class}
\alias{request_lipidbank_lipid_class}
\title{Request Lipidbank database}
\usage{
request_lipidbank_lipid_class(
  url = "https://lipidbank.jp/",
  lipid_class = c("All data", "Acylglycerol", "Bile Acid", "Fatty acid",
    "Long chain alcohol", "Long chain aldehyde", "Long chain base and Ceramide",
    "Eicosanoid", "Ether type lipid", "Carotenoid", "Coenzyme Q", "Vitamin A",
    "Vitamin D", "Vitamin E", "Vitamin F", "Vitamin K", "Glycosphingolipid",
    "Glycoglycerolipid and others", "Isoprenoid", "Lipid peroxide", "Lipoamino acid",
    "Lipopolysaccharide", "Lipoprotein", "Mycolic acid", "Glycerophospholipid",
    "Sphingophospholipid", "Steroid", "Wax")
)
}
\arguments{
\item{url}{Default is "https://hmdb.ca/metabolites".}

\item{lipid_class}{lipid_class}
}
\value{
A data frame or list.
}
\description{
Request one specific the metabolite information in HMDB
}
\examples{
x = request_lipidbank_lipid_class(lipid_class = "Wax")
head(x)
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
