% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13_LIPIDMAPS.R
\name{request_lipidmaps_lipid}
\alias{request_lipidmaps_lipid}
\title{Request one lipid information in LIPIDMAPS}
\usage{
request_lipidmaps_lipid(
  url = "https://www.lipidmaps.org/databases/lmsd",
  lipid_id = "LMFA01030001"
)
}
\arguments{
\item{url}{Default is "https://www.lipidmaps.org/databases/lmsd".}

\item{lipid_id}{lipid id. For example, LMFA01030001}
}
\value{
A data frame.
}
\description{
Request one lipid information in LIPIDMAPS
}
\examples{
request_lipidmaps_lipid(lipid_id = "LMFA01030001")
request_lipidmaps_lipid(lipid_id = "LMGL00000140")
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
