% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15_PUBCHEM.R
\name{request_pubchem_compound}
\alias{request_pubchem_compound}
\title{Request the information of one compound from PubChem}
\usage{
request_pubchem_compound(compound_id = "5288826")
}
\arguments{
\item{compound_id}{compound_id name}
}
\value{
A list.
}
\description{
Request the information of one compound from PubChem
}
\examples{
x <- request_pubchem_compound(compound_id = "5288826")
names(x)
x$Computed_Descriptors
x$Molecular_Formula
x$Other_Identifiers
x$Synonyms
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
