% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/19_REACTOME.R
\name{request_reactome_organisms_info}
\alias{request_reactome_organisms_info}
\title{Retrieve Available Organisms in Reactome}
\usage{
request_reactome_organisms_info()
}
\value{
A sorted vector of organism names.
}
\description{
This function retrieves the list of organisms available in the Reactome database.
}
\examples{
organisms <- request_reactome_organisms_info()
print(organisms)
}
\author{
Xiaotao Shen (\email{xiaotao.shen@outlook.com})
}
