% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/19_REACTOME.R
\name{request_reactome_pathway}
\alias{request_reactome_pathway}
\title{Retrieve Detailed Reactome Pathway Information}
\usage{
request_reactome_pathway(pathway_id = "R-HSA-5652084")
}
\arguments{
\item{pathway_id}{Character. The Reactome pathway ID (e.g., "R-HSA-5652084").}
}
\value{
A list containing pathway details, including name, description, and involved metabolites/proteins.
}
\description{
This function fetches and parses a given Reactome pathway in SBML format.
}
\examples{
pathway_info <- request_reactome_pathway("R-HSA-5652084")
print(pathway_info$pathway_name)
print(pathway_info$pathway_description)
}
\author{
Xiaotao Shen (\email{xiaotao.shen@outlook.com})
}
