% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/19_REACTOME.R
\name{request_reactome_pathway_info}
\alias{request_reactome_pathway_info}
\title{Retrieve Reactome Pathways for a Given Organism}
\usage{
request_reactome_pathway_info(organism = "Homo sapiens")
}
\arguments{
\item{organism}{Character. The name of the organism. Default is "Homo sapiens".}
}
\value{
A data frame containing pathway ID, pathway name, and organism.
}
\description{
This function retrieves pathway information for a specified organism from Reactome.
}
\examples{
pathways <- request_reactome_pathway_info("Homo sapiens")
head(pathways)
}
\author{
Xiaotao Shen (\email{xiaotao.shen@outlook.com})
}
