% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/19_REACTOME.R
\name{request_reactome_reaction}
\alias{request_reactome_reaction}
\title{Request one specific reaction information in Reactome}
\usage{
request_reactome_reaction(reaction_id = "R-HSA-8876188")
}
\arguments{
\item{reaction_id}{reaction id. For example, reaction_id}
}
\value{
Alist.
}
\description{
Request one specific reaction information in Reactome
}
\examples{
x = request_reactome_reaction(reaction_id = "R-HSA-8876188")
x$reactants
x$products
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
