% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/21_WIKIDATA.R
\name{request_wikidata_metabolite}
\alias{request_wikidata_metabolite}
\title{Retrieve Metabolite Information from Wikidata}
\usage{
request_wikidata_metabolite(metabolite_id = "Q18216")
}
\arguments{
\item{metabolite_id}{A character string specifying the Wikidata ID of the metabolite (default: "Q18216").}
}
\value{
A list containing detailed metabolite information, including:
  \itemize{
    \item \code{metabolite_id} - The Wikidata ID of the metabolite.
    \item \code{metabolite_name} - The name of the metabolite.
    \item \code{metabolite_description} - A brief description of the metabolite.
    \item \code{aliases} - Alternative names for the metabolite.
    \item \code{wikipedia_url} - The Wikipedia URL for the metabolite.
    \item \code{chemical_formula} - The chemical formula of the metabolite.
    \item \code{smiles} - The SMILES notation of the metabolite.
    \item \code{inchi} - The InChI representation of the metabolite.
    \item \code{inchikey} - The InChIKey of the metabolite.
    \item \code{cas_id} - The CAS Registry Number.
    \item \code{pubchem_cid} - The PubChem Compound ID.
    \item \code{chebi_id} - The ChEBI ID.
    \item \code{chemspider_id} - The ChemSpider ID.
    \item \code{chembl_id} - The ChEMBL ID.
    \item \code{kegg_id} - The KEGG Compound ID.
    \item \code{hmdb_id} - The Human Metabolome Database (HMDB) ID.
    \item \code{lipidmaps_id} - The LIPID MAPS ID.
    \item \code{drugbank_id} - The DrugBank ID.
    \item \code{rxnorm_id} - The RxNorm ID.
    \item \code{ndf_rt_id} - The NDF-RT ID.
    \item \code{mesh_descriptor_id} - The MeSH Descriptor ID.
    \item \code{unii} - The Unique Ingredient Identifier (UNII).
    \item \code{guide_to_pharmacology_ligand_id} - The Guide to Pharmacology Ligand ID.
    \item \code{atc_code} - The Anatomical Therapeutic Chemical (ATC) classification code.
    \item \code{echa_substance_infocard_id} - The ECHA Substance Infocard ID.
    \item \code{chembl_id} - The ChEMBL ID.
    \item \code{freebase_id} - The Freebase ID.
    \item \code{reaxys} - The Reaxys Registry Number.
    \item \code{zvg_number} - The ZVG number.
    \item \code{probes_and_drugs_id} - The Probes and Drugs ID.
    \item \code{australian_register_of_therapeutic_goods_id} - The Australian Register of Therapeutic Goods ID.
    \item \code{other_identifiers} - Various other database identifiers (e.g., SPLASH, DSSTox, PDB Ligand ID, etc.).
  }
}
\description{
This function queries Wikidata for detailed information about a given metabolite using its Wikidata ID.
}
\examples{
request_wikidata_metabolite("Q18216") # Fetches metabolite data for a given Wikidata ID.
}
