% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20_WIKIPATHWAYS.R
\name{request_wikipathway}
\alias{request_wikipathway}
\title{Retrieve Detailed Pathway Data from WikiPathways}
\usage{
request_wikipathway(pathway_id = "WP5293")
}
\arguments{
\item{pathway_id}{A character string specifying the WikiPathways ID (e.g., "WP5293").}
}
\value{
A list containing:
  \itemize{
    \item \code{pathway_info} - Metadata about the pathway.
    \item \code{pathway_description} - A description of the pathway.
    \item \code{node_info} - A data frame containing gene/metabolite nodes.
    \item \code{edge_info} - A data frame containing interaction edges.
  }
}
\description{
This function fetches detailed information about a specific pathway from WikiPathways using its pathway ID.
}
\examples{
data = request_wikipathway("WP5293")
data$pathway_info
}
