% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20_WIKIPATHWAYS.R
\name{request_wikipathway_info}
\alias{request_wikipathway_info}
\title{Retrieve Pathway Information for a Specific Organism}
\usage{
request_wikipathway_info(organism = "Homo sapiens")
}
\arguments{
\item{organism}{A character string specifying the organism name. Default is "Homo sapiens".}
}
\value{
A data frame containing pathway information, including pathway names, IDs, and species.
}
\description{
This function fetches all available pathways from WikiPathways for a specified organism.
}
\examples{
request_wikipathway_info("Homo sapiens")
}
