% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20_WIKIPATHWAYS.R
\name{request_wikipathway_organisms_info}
\alias{request_wikipathway_organisms_info}
\title{Retrieve Available Organisms from WikiPathways}
\usage{
request_wikipathway_organisms_info()
}
\value{
A character vector of organism names available in the WikiPathways database.
}
\description{
This function queries WikiPathways to obtain a list of organisms for which pathways are available.
}
\examples{
request_wikipathway_organisms_info()
}
