% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15_WIKIPEDIA.R
\name{request_wikipedia_compound}
\alias{request_wikipedia_compound}
\title{Request the information of one compound}
\usage{
request_wikipedia_compound(compound_id = "Microcystin-LR")
}
\arguments{
\item{compound_id}{compound_id name}
}
\value{
A data frame.
}
\description{
Request the information of one compound
}
\examples{
request_wikipedia_compound(compound_id = "Microcystin-LR")
request_wikipedia_compound(compound_id = "Glucose")
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
