% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WIKIPEDIA.R
\name{request_wikipedia_scientific_classification}
\alias{request_wikipedia_scientific_classification}
\title{Request the scientific classification of one species from wikipedia}
\usage{
request_wikipedia_scientific_classification(species_id = "Aaptos ciliata")
}
\arguments{
\item{species_id}{species_id}
}
\value{
A data frame.
}
\description{
Request the scientific classification of one species from wikipedia
}
\examples{
x = request_wikipedia_scientific_classification(species_id = "Aaptos ciliata")
x
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
