% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_HMDB.R
\name{search_hmdb_database}
\alias{search_hmdb_database}
\title{Search metabolite in HMDB}
\usage{
search_hmdb_database(
  name = "5-methoxysalicylic acid",
  mz = 168.042625,
  data_database,
  mz_error_cutoff = 100,
  similarity_score_cutoff = 0.5
)
}
\arguments{
\item{name}{metabolite name}

\item{mz}{mz}

\item{data_database}{data_database, metID format}

\item{mz_error_cutoff}{mz_error_cutoff}

\item{similarity_score_cutoff}{similarity_score_cutoff}
}
\value{
a data frmae
}
\description{
Search metabolite in HMDB
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
