% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_utils.R
\name{show_progresser}
\alias{show_progresser}
\title{show_progresser}
\usage{
show_progresser(
  index = 1:1000,
  progresser = c(1, 10, 20, 30, 40, 50, 60, 70, 80, 90, 100)
)
}
\arguments{
\item{index}{index for loop}

\item{progresser}{progresser}
}
\value{
A data.frame
}
\description{
show_progresser
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
