% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_utils.R, R/utils.R
\name{massdatabase_packages}
\alias{massdatabase_packages}
\title{List all packages in the massdatabase}
\usage{
massdatabase_packages(include_self = TRUE)

massdatabase_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include massdatabase in the list?}
}
\value{
massdatabase_packages

massdatabase packages
}
\description{
List all packages in the massdatabase

List all packages in the massdatabase
}
\examples{
massdatabase_packages()
massdatabase_packages()
}
