% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_BIGG.R
\name{request_bigg_version}
\alias{request_bigg_version}
\title{Request BIGG version}
\usage{
request_bigg_version(url = "http://bigg.ucsd.edu/api/v2/database_version")
}
\arguments{
\item{url}{Default is "https://foodb.ca/compounds".}
}
\value{
version.
}
\description{
Request BIGG version
}
\examples{
request_bigg_version()
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
