% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FOODB.R
\name{request_foodb_food_info}
\alias{request_foodb_food_info}
\title{Request all the food information in FoodB}
\usage{
request_foodb_food_info(
  url = "https://foodb.ca/foods",
  sleep = 1,
  pages = c(1:32)
)
}
\arguments{
\item{url}{Default is "https://foodb.ca/foods".}

\item{sleep}{Default is 1 second.}

\item{pages}{default is from 1:2838}
}
\value{
A data frame.
}
\description{
Request all the food information in FoodB
}
\examples{
x = request_foodb_food_info(pages = 1)
head(x)
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
