% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_data.R
\docType{data}
\name{smpdb_primary_pathway_id}
\alias{smpdb_primary_pathway_id}
\title{smpdb_primary_pathway_id}
\description{
smpdb_primary_pathway_id
}
\examples{
data(smpdb_primary_pathway_id)
}
\keyword{data}
